Rekam dan ubah registrasi by wahid issue 23

+++ trunk/Apdol/src/java/apdol/entity/User.java
@@ -40,7 +40,7 @@ public class User implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Column(name = "username")
     private String username;
     @Column(name = "password")
@@ -60,15 +60,15 @@ public class User implements Serializable {
     public User() {
     }
 
-    public User(Integer id) {
+    public User(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
     
@@ -146,4 +146,7 @@ public class User implements Serializable {
     public String toString() {
         return "apdol.entity.User[ id=" + id + " ]";
     }
-}
+
+
+ }
+
